+++ packages/plus/ai/src/providers/openAICompatibleProviderBase.ts
@@ -90,7 +90,6 @@ export abstract class OpenAICompatibleProviderBase<T extends AIProviders> implem
 			scope?.error(ex, `Unable to ${getActionName(action)}: (${model.provider.name})`);
 			if (ex instanceof AIError) throw ex;
 
-			debugger;
 			throw new Error(`Unable to ${getActionName(action)}: (${model.provider.name}) ${ex.message}`, {
 				cause: ex,
 			});